### 2022-10-09 #4832 @editorialbot
[PRE REVIEW]: Metagenomic classification with KrakenUniq on low-memory computers
**Submitting author:** @salzberg (<a href="http://orcid.org/0000-0002-8859-7432">Steven Salzberg</a>)
**Repository:** https://github.com/fbreitwieser/krakenuniq
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
