+++ paper.bib
@@ -6,6 +6,7 @@
   number={3},
   pages={1--12},
   year={2014},
+  doi={10.1186/gb-2014-15-3-r46},
   publisher={BioMed Central}
 }
 
@@ -17,6 +18,7 @@
   number={1},
   pages={1--10},
   year={2018},
+  doi={10.1186/s13059-018-1568-0},
   publisher={BioMed Central}
 }
 
@@ -28,6 +30,7 @@
   number={1},
   pages={1--13},
   year={2019},
+  doi={10.1186/s13059-019-1891-0},
   publisher={Springer}
 }
 
@@ -39,6 +42,7 @@
   number={6},
   pages={e1006277},
   year={2018},
+  doi={10.1371/journal.pcbi.1006277},
   publisher={Public Library of Science San Francisco, CA USA}
 }
 
