### 2022-10-23 #4832 @salzberg
@lpantano, I think what that is asking is for me to add the DOI to the bibtex file. The bibtex file was already formatted correctly and all references were there, but I just added DOIs to all of those references. Please let me know if you're asking for something else.
