### 2022-11-10 #4908 @MaximLippeveld
Hi @lpantano

The submitter only has 5 commits on the repository. I assume this is not enough to meet the "Contribution and authorship" check.

I am also not sure about the scholarly effort. The paper describes only one added feature to an existing software. While I am convinced that the added feature is useful, I'm not sure if one feature is substantial enough to warrant a paper.

Could you weigh in on this?
