### 2022-11-10 #4908 @salzberg
> Hi @lpantano
>
> The submitter only has 5 commits on the repository. I assume this is not enough to meet the "Contribution and authorship" check.
>
> I am also not sure about the scholarly effort. The paper describes only one added feature to an existing software. While I am convinced that the added feature is useful, I'm not sure if one feature is substantial enough to warrant a paper.
>
> Could you weigh in on this?

I am one of the authors. The new feature required substantial re-engineering of the code, and it provides a really, really useful new capability. This software uses a custom database that is huge, typically around 400 GB, and the database must be in RAM for the original version to work. That limits users to finding a server with more than 400 GB of RAM, and for many users they simply can't do it. My own colleagues in the School of Medicine were recently scrambling (unsuccessfully) to find funds to buy a new, large-memory server so they could run this software (KrakenUniq) locally. They are using it for medical data so they can't simply run things on an external cloud server like Amazon.
   The new "feature" is an elegant way to read in the database in chunks, process all of the other input data, store the partial results in temporary files, and then read in the next chunk. The user can divide the data into as many chunks as they like, and the runtime penalty is really surprisingly modest.
   This work was done by a postdoc, Christopher Pockrandt (1st author), who was not part of the original papers describing this system. How else can he get credit if this isn't considered "substantial enough to warrant a paper"? I hope you'll agree that we need to make it possible for scientists to publish good software engineering work - otherwise, there is little incentive to maintain research software.
-Steven Salzberg
