### 2022-11-10 #4908 @MaximLippeveld
Hi Steven,

Thanks for your clarifications. I definitely agree we need to publish good software engineering work; I'm reviewing for JOSS after all :slightly_smiling_face:

I do feel that the complexity of the proposed solution as you describe it now, is not reflected in the paper. Maybe the first paragraph of the section "Database chunking" can be elaborated on? For example, is there anything specific about merging the k-mer lookups from the different chunks? How does the chunked reading from disk work? Adding pseudo code to describe the algorithm could help as well.
