### 2022-11-10 #4908 @salzberg
I'll let Christopher know and see if we can add some more details about the temporary data structure and how we merge them.
