### 2022-11-13 #4908 @Jessime
Hi, everyone!

First, I want to quickly caveat that this is my first review, Github based or otherwise. I'm looking forward to getting into it, and hope you'll bear with me as I figure out all of the specifics.

I'm holding out hope that I'll be able to get up and running with a `docker pull`, but in the meantime, here's a list of questions/concerns/suggestions based on a read through of the paper/documentation/git history.

* Re: **License**, I've checked the box, but I'm curious about the filename choice. It's rare to see anything other than `LICENSE` or perhaps `LICENSE.md`. Does JOSS have a convention here?
* Re: **Contribution and authorship**, it looks like this meets the criteria of three months of work, and I'm a huge proponent of smaller papers and review cycles. Though, I will note that understanding what work was accomplished is made substantially more difficult by the somewhat convoluted git history and repo forking. I don't think there's much to be done about the present state of things, but it's worth pointing out that attempting to follow git best practices will certain yield real-world improvements in software quality in the future.
* Re: **Installation**, I'd like to strongly advocate for a `Dockerfile` to be included in the repo, and if at all possible, corresponding Docker and/or Singularity images to be made available. Virtually every organization using KrakenUniq at scale will wrap this repo in an image. Having it provide by the repo maintainers will remove a significant amount of unnecessary re-work. _Note:_ Conda is also a reasonable package manager, but is more niche outside of academia.
* Re: **Performance**, The paper only shows a single performance comparison, but also notes that "Running times and speedups can vary significantly depending on the type of storage (e.g., databases on network storage can take longer to load)". Given the prevalence of networked storage, can we see what performance looks like in that case? More generally, it's likely helpful to be explicit about "worst-case" scenarios. Are there other common hardware considerations that have significant negative impact on runtime?
* Re: **Example usage**, The overall example usage for KrakenUniq seems pretty reasonable. For this new feature, however, it isn't clear how the `--preload-size` and `--preload` flags interact. Will `--preload-size` work without `--preload` or does it imply `--preload-size`? In fact, I believe the cleaner API is to not have a new flag. `--preload` can just be passed either with a size `--preload 8GB`, or without `--preload`, the latter implying preloading everything. Keeping a tight list of flags a user needs to remember makes for a better user experience.
* Re: **Functionality documentation**, Similar to the point above, an explicit section of `Options` is often the most important section of a README. [`fd` is a great example](https://github.com/sharkdp/fd#command-line-options). Or, in other words, [this function](https://github.com/fbreitwieser/krakenuniq/blob/25dccf4035b4e5ecb3593b32fc2d341e7f2bd74a/src/classify.cpp#L1164) should be in the README.
* Re: **Community guidelines**, It seems like a brief `CONTRIBUTING.md` file would get this box checked. There are mainly guidelines available, [like this one](https://mozillascience.github.io/working-open-workshop/contributing/)
