+++ paper.md
@@ -105,10 +105,9 @@ a binary search on the minimizer table to find the largest minimizer such that t
 the corresponding chunk of the *k-mer* table together use not more than the specified amount of memory. It then loads the chunk
 of the minimizer table and the corresponding chunk of the k-mer table and iterates
 over all of the reads provided as input.  The code looks up all *k-mers* in the reads in the loaded chunk of the database.
-The taxonomical ids for the *k-mers* that matched a k-mer in the chunk (or placeholders for *k-mers* without a hit) are stored in a temporary file on disk.
-With every chunk iteration taxonomical ids of newly found *k-mers* are added to the file. This process is repeated until
-the entire database has been processed. The *k-mer* lookups stored in the temporary file are then used to classify the reads.
-Classification results will be identical to running in the default mode; i.e., database chunking does not alter the output.
+The taxonomical ids for the *k-mers* that matched a k-mer in the chunk (or placeholders for *k-mers* without a hit) are stored in a temporary file on disk for each read.
+With every chunk iteration taxonomical ids of newly found *k-mers* in the reads are added to the temporary file. This process is repeated until
+the entire database has been processed. At this point the temporary file contains classification for all k-mers in the reads. KrakenUniq then reads the temporary file, collects *k-mer* hits for every read and uses them to classify each read. Classification results will be identical to running in the default mode; i.e., database chunking does not alter the output.
 
 This new feature makes it feasible to run KrakenUniq on very large datasets and huge databases on
 virtually any computer, even a laptop, while providing exact classifications that are identical to those
