### 2022-11-28 6d9879d
Create CONTRIBUTING.md
