### 2022-11-28 6ec15a2
Create CODE_OF_CONDUCT.md
