+++ CONTRIBUTING.md
@@ -1,81 +1,131 @@
-# Contributing
+<!-- omit in toc -->
+# Contributing to KrakenUniq
 
-We welcome contributions in the forms of helpful suggesions, code or issues reported by github issue tracker.
-In reporting an issue please include enough information (full command line, link to downloadable data file(s)) for us to be able to reproduce the issue.  
+First off, thanks for taking the time to contribute! ❤️
 
-Please note we have a code of conduct, please follow it in all your interactions with the project.
+All types of contributions are encouraged and valued. See the [Table of Contents](#table-of-contents) for different ways to help and details about how this project handles them. Please make sure to read the relevant section before making your contribution. It will make it a lot easier for us maintainers and smooth out the experience for all involved. The community looks forward to your contributions. 🎉
+
+> And if you like the project, but just don't have time to contribute, that's fine. There are other easy ways to support the project and show your appreciation, which we would also be very happy about:
+> - Star the project
+> - Tweet about it
+> - Refer this project in your project's readme
+> - Mention the project at local meetups and tell your friends/colleagues
+
+<!-- omit in toc -->
+## Table of Contents
+
+- [Code of Conduct](#code-of-conduct)
+- [I Have a Question](#i-have-a-question)
+- [I Want To Contribute](#i-want-to-contribute)
+- [Reporting Bugs](#reporting-bugs)
+- [Suggesting Enhancements](#suggesting-enhancements)
+
 
 ## Code of Conduct
 
-### Our Pledge
+This project and everyone participating in it is governed by the
+[KrakenUniq Code of Conduct](https://github.com/fbreitwieser/krakenuniqblob/master/CODE_OF_CONDUCT.md).
+By participating, you are expected to uphold this code. Please report unacceptable behavior
+to .
 
-In the interest of fostering an open and welcoming environment, we as
-contributors and maintainers pledge to making participation in our project and
-our community a harassment-free experience for everyone, regardless of age, body
-size, disability, ethnicity, gender identity and expression, level of experience,
-nationality, personal appearance, race, religion, or sexual identity and
-orientation.
 
-### Our Standards
+## I Have a Question
 
-Examples of behavior that contributes to creating a positive environment
-include:
+> If you want to ask a question, we assume that you have read the available [Documentation](https://github.com/fbreitwieser/krakenuniq).
 
-* Using welcoming and inclusive language
-* Being respectful of differing viewpoints and experiences
-* Gracefully accepting constructive criticism
-* Focusing on what is best for the community
-* Showing empathy towards other community members
+Before you ask a question, it is best to search for existing [Issues](https://github.com/fbreitwieser/krakenuniq/issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in this issue. It is also advisable to search the internet for answers first.
 
-Examples of unacceptable behavior by participants include:
+If you then still feel the need to ask a question and need clarification, we recommend the following:
 
-* The use of sexualized language or imagery and unwelcome sexual attention or
-advances
-* Trolling, insulting/derogatory comments, and personal or political attacks
-* Public or private harassment
-* Publishing others' private information, such as a physical or electronic
-  address, without explicit permission
-* Other conduct which could reasonably be considered inappropriate in a
-  professional setting
+- Open an [Issue](https://github.com/fbreitwieser/krakenuniq/issues/new).
+- Provide as much context as you can about what you're running into.
+- Provide project and platform versions (nodejs, npm, etc), depending on what seems relevant.
 
-### Our Responsibilities
+We will then take care of the issue as soon as possible.
 
-Project maintainers are responsible for clarifying the standards of acceptable
-behavior and are expected to take appropriate and fair corrective action in
-response to any instances of unacceptable behavior.
+<!--
[…]
