### 2022-11-28 #4908 @alekseyzimin
Thank you @MaximLippeveld and @Jessime for your reviews!

@MaximLippeveld , we added more detail as to how the modification works at the beginning of "Database chunking" section:
"Under this new algorithm, KrakenUniq loads one chunk of the database into memory at a time. KrakenUniq performs
a binary search on the minimizer table to find the largest minimizer such that the chunk of the minimizer table and
the corresponding chunk of the *k-mer* table together use not more than the specified amount of memory. It then loads the chunk
of the minimizer table and the corresponding chunk of the k-mer table and iterates
over all of the reads provided as input.  The code looks up all *k-mers* in the reads in the currently-loaded chunk of the database.
The taxonomic IDs for the *k-mers* that matched a k-mer in the chunk (or placeholders for *k-mers* without a hit) are stored in a temporary file on disk for each read.
With every chunk iteration, taxonomic IDs of newly found *k-mers* in the reads are appended to the temporary file. This process is repeated until
the entire database has been processed. At this point the temporary file contains taxonomic IDs for all k-mers in the reads that matched any part of the database. KrakenUniq then reads the temporary file, collects *k-mer* hits for every read, and uses them to classify each read. Classification results will be identical to running in the default mode; i.e., database chunking does not alter the output."

@Jessime, here are our responses to your review:

**License:** we renamed the LICENSE.GPL-3.0 to LICENSE.
Contribution and Authorship: several students and postdocs have contributed to KrakenUniq sequentially. Once a student or postdoc graduates, they sometimes stop working on the project, and maintenance and improvements have to be picked up by another person.  That explains the git history.

**Installation:** At this time KrakenUniq is available through github and BioConda. We will definitely consider adding a docker image once we have manpower to do it.

**Performance:** The default behavior (without any preload) the worst one can do.  We modified the paragraph describing the performance as follows:
"Running times can vary significantly depending on the type of storage (e.g., databases
on network storage can take longer to load) and the size of the read dataset (i.e., classifying a small
number of reads does not justify preloading the entire database, especially of fast storage). The speed of loading
the database is not impacted by the `--preload-size` option because the database is still read in a linear way.
Saving intermediate files from the chunks is done in the same way as in the original code.
The only difference is that now classification results from all individual chunks are concatenated into a single file, which is read once all chunks are finished. **Table 1** shows that in a typical use case, even when the database does fit in RAM, loading the entire database (`--preload` option) is far faster than memory mapping (14 minutes versus 48 hours). Loading the database by chunks adds overhead because of the need to
iterate over the reads multiple times, but is still comparable to pre-loading the entire database and
highly recommended when not enough main memory is available. For example, limiting the database
to 8G, which means it can be loaded even on a standard laptop computer, increased the running
time only about 3.4-fold, even though the database was broken into 49 chunks. For large read datasets we expect that setting the `--preload` or `--preload-size` flag will always be faster than the default behavior of memory mapping. The format of the
databases used by the new algorithm has not changed, hence all previously built databases for Kraken
and KrakenUniq can be used."

**Example usage:**  We would like to keep the --preload flag for compatibility with legacy scripts.  We added "overrides --preload flag" to the usage:
--preload-size SIZE     Loads DB into memory in chunks of SIZE, e.g. 500M or 7G (if RAM is small), overrides --preload flag

**Functionality documentation:** We added the usage statement to the README.md, in a similar way it was done for the `fd` example.

**Community guidelines:** We added CONTRIBUTING.md and CODE_OF_CONDUCT.md to the repository.
