### 2022-11-29 #4908 @MaximLippeveld
Hi, thanks for the updates. The "database chunking" section is now very clear for me!

A couple more remarks:
- Installation: I had to add change the installation command to `conda install -c bioconda krakenuniq`. The repo's README doesn't include `-c bioconda` at the moment.
- Statement of need: I would like to see a short extra paragraph on the need for the specific functionality that is being introduced in this paper.
- Reproducibility: Could you provide a script for reproducing Table 1? (Or point me to it, if I've missed it.)
- References: Please add a reference for [GenBank](https://pubmed.ncbi.nlm.nih.gov/23193287/). Perhaps the [minimizers](https://academic.oup.com/bioinformatics/article/20/18/3363/202143) also needs a reference.
