### 2022-11-29 891b64b
Update README.md
added "-c bioconda" to the instruction on how to install with conda.
