+++ paper.bib
@@ -1,3 +1,27 @@
+@article{benson2012genbank,
+  title={GenBank},
+  author={Benson, Dennis A and Cavanaugh, Mark and Clark, Karen and Karsch-Mizrachi, Ilene and Lipman, David J and Ostell, James and Sayers, Eric W},
+  journal={Nucleic acids research},
+  volume={41},
+  number={D1},
+  pages={D36--D42},
+  year={2012},
+  doi={10.1093/nar/gks1195},
+  publisher={Oxford University Press}
+}
+
+@article{roberts2004reducing,
+  title={Reducing storage requirements for biological sequence comparison},
+  author={Roberts, Michael and Hayes, Wayne and Hunt, Brian R and Mount, Stephen M and Yorke, James A},
+  journal={Bioinformatics},
+  volume={20},
+  number={18},
+  pages={3363--3369},
+  year={2004},
+  doi={10.1093/bioinformatics/bth408},
+  publisher={Oxford University Press}
+}
+
 @article{wood2014kraken,
   title={Kraken: ultrafast metagenomic sequence classification using exact alignments},
   author={Wood, Derrick E. and Salzberg, Steven L.},
