### 2022-11-30 #4908 @MaximLippeveld
Great!

Could you also add instructions on downloading `SRR12486990.fastq`?
