### 2022-11-30 45ee581
Update paper.md
added reference to NCBI site where one can download the data in Table 1, https://www.ncbi.nlm.nih.gov/sra/SRR12486990.
