### 2022-11-30 #4908 @salzberg
> Great!
>
> Could you also add instructions on downloading `SRR12486990.fastq`?

Yes, just did this. That's an NCBI accession, and we've put the link in the Table caption, pointing to https://www.ncbi.nlm.nih.gov/sra/SRR12486990.
