### 2022-12-02 d6de665
Update paper.md
Added cite to another NIH grant.
