### 2022-12-05 #4908 @MaximLippeveld
Thanks for the updates. I was able to run the command successfully in 44min using `--preload-size 16G` and `--threads 10` on an Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz. Due to computational constraints, I am not able to reproduce all timing results, unfortunately.

@lpantano My review is completed!
