### 2022-12-13 #4908 @Jessime
Hey everyone, things are looking good. I just have one more point on the **Installation**.
Namely, adding the following `Dockerfile` will allow users who don't have access to conda to be able to use your tool:

```Dockerfile
FROM alpine:latest

RUN apk add --no-cache \
  bash \
  perl \
  make \
  g++ \
  bzip2-dev \
  zlib-dev

WORKDIR /app
COPY . .
RUN ./install_krakenuniq.sh /usr/local/bin

CMD ["bash"]
```

It's then a matter of:

```bash
docker build -t krakenuniq .
docker run --rm -it -it krakenuniq krakenuniq --help
```

to install and verify things look good. Would it be possible to add the `Dockerfile` and the instructions?
