### 2022-12-13 6623792
Added instructions for Docker per @Jessime suggestion
