### 2022-12-18 05093d0
build_db.sh: find/use gzipped files, avoid erroneous warnings when shrinking db
