+++ README.md
@@ -8,6 +8,10 @@ False-positive identifications are a significant problem in metagenomics classif
 We now have two standard Kraken1/KrakenUniq databases available for free download from the Amazon cloud. You can find links at [https://benlangmead.github.io/aws-indexes/k2](https://benlangmead.github.io/aws-indexes/k2). One is our "standard" database with all RefSeq bacteria, archaea, and viruses, plus common vectors and the human genome. The other is all of the first database plus all available genomes of eukaryotic pathogens. Each DB is over 300GB, and by downloading them you can avoid having to build them yourself.
 
 # Announcements
+
+## New release v1.0.3
+In this release we added a Dockerfile (thanks @Jessime).  There are also few minor updates to documentation. This code of this release has been reviewed in connection with our publication in JOSS.
+
 ## New release v1.0.2
 This release fixes the issue with possibly incorrect output produced when running multiple krakenuniq processes in the same folder with --paired input files.  
 
