### 2022-12-19 #4908 @alekseyzimin
Hello @lpantano, thank you for letting us know about the next steps.  We performed the requested updates, please see below.
Make a tagged release of your software, and list the version tag of the archived version here.
Response: released v1.0.3
Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID. Please list the DOI of the archived version here.
Response: 10.5281/zenodo.7459217
