### 2022-12-28 #4908 @Kevin-Mattheus-Moerman
@alekseyzimin I am the AEiC on this track, and here to help process this work for acceptance in JOSS. I have reviewed the paper and archive and most seems in order, so we can proceed with acceptance. I do have one point below which you should fix at your earliest convenience:

- [ ] Please modify the license listed on your ZENODO archive. It currently refers to a creative commons license, but this should be your software license: GPL-3.0 license.

I will now proceed with to acceptance in JOSS.
